Two minor fixes for deploying to bare metal
Summary:
* remove half.tab from csproj (has never been used)
* check for existence before removing the test database

Test Plan: unit tests

Reviewers: tberman, stephan, justin

Reviewed By: tberman

CC: serdar, stephan, justin, tberman

Differential Revision: http://phabricator.thefactory.com/D1635

+++ csharp/TheFactory.DatastoreTests/src/ScaleTests.cs
@@ -33,7 +33,9 @@ namespace TheFactory.DatastoreTests {
 
         void TestRoundTrip(Options opts, int numWriters, IEnumerable<IKeyValuePair> goldenKVs) {
             var dbPath = Path.Combine(Path.GetTempPath(), "db");
-            Directory.Delete(dbPath, true);
+            if (Directory.Exists(dbPath)) {
+                Directory.Delete(dbPath, true);
+            }
 
             var taskOptions = new ParallelOptions();
             taskOptions.MaxDegreeOfParallelism = numWriters;
